311223 1332 spinner modified
